move components into separate files  (#1190)
